requirements.txt
Flask
gunicorn
pillow
flask_cors
pytesseract